+++ card-game-service/src/main/java/jrh/game/service/lobby/MatchQueue.java
@@ -4,14 +4,23 @@ import jrh.game.common.account.AccountId;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
+import java.util.Map;
 import java.util.concurrent.BlockingDeque;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.Executors;
+import java.util.concurrent.Future;
 import java.util.concurrent.LinkedBlockingDeque;
+import java.util.concurrent.ScheduledExecutorService;
+import java.util.concurrent.TimeUnit;
 
 public class MatchQueue {
 
     private static final Logger logger = LogManager.getLogger(MatchQueue.class);
+    private static final long QUEUE_TIMEOUT_MS = 20 * 1000;
 
     private final BlockingDeque<AccountId> queue = new LinkedBlockingDeque<>();
+    private final ScheduledExecutorService scheduledExecutorService = Executors.newSingleThreadScheduledExecutor();
+    private final Map<AccountId, Future<?>> leaveQueueFutures = new ConcurrentHashMap<>();
 
     public int size() {
         return queue.size();
@@ -23,6 +32,7 @@ public class MatchQueue {
             queue.addLast(accountId);
         }
         logger.debug("Queue now contains {} accounts", queue.size());
+        refresh(accountId);
     }
 
     public void prioritise(AccountId accountId) {
@@ -30,6 +40,7 @@ public class MatchQueue {
         logger.info("accountId={} joined the front of the queue", accountId);
         queue.addFirst(accountId);
         logger.debug("Queue now contains {} accounts", queue.size());
+        refresh(accountId);
     }
 
     public void remove(AccountId accountId) {
@@ -37,10 +48,26 @@ public class MatchQueue {
             logger.info("accountId={} left the queue", accountId);
         }
         logger.debug("Queue now contains {} accounts", queue.size());
+        Future<?> leaveQueue = leaveQueueFutures.remove(accountId);
+        if (leaveQueue != null) {
+            leaveQueue.cancel(true);
+        }
     }
 
     public boolean contains(AccountId accountId) {
         return queue.contains(accountId);
+    }
+
+    public void refresh(AccountId accountId) {
+        Future<?> leaveQueue = leaveQueueFutures.remove(accountId);
+        if (leaveQueue != null) {
+            leaveQueue.cancel(true);
+        }
+        Future<?> newLeaveQueue = scheduledExecutorService.schedule(() -> {
+            remove(accountId);
+            leaveQueueFutures.remove(accountId);
+        }, QUEUE_TIMEOUT_MS, TimeUnit.MILLISECONDS);
+        leaveQueueFutures.put(accountId, newLeaveQueue);
     }
 
     public AccountId poll() {